# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para generar y consultar facturas CFDI usando Facturama",
        "title": "Facturama API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Emmanuel Rdz",
            "email": "emmanuelrdz@example.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/cfdi": {
            "get": {
                "description": "Devuelve una lista de facturas emitidas con filtros",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CFDI"
                ],
                "summary": "Consultar CFDIs emitidos",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tipo de CFDI (issued)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Folio inicial",
                        "name": "folioStart",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Folio final",
                        "name": "folioEnd",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC",
                        "name": "rfc",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Nombre del receptor",
                        "name": "taxEntityName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Fecha inicio (dd/mm/yyyy)",
                        "name": "dateStart",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Fecha fin (dd/mm/yyyy)",
                        "name": "dateEnd",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Estado de factura",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Ordenar por folio",
                        "name": "orderNumber",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Número de página",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Genera una nueva factura CFDI (global o normal)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CFDI"
                ],
                "summary": "Crear CFDI",
                "parameters": [
                    {
                        "description": "Datos del CFDI (estructura completa)",
                        "name": "factura",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CfdiRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/cfdi/{id}/download": {
            "get": {
                "description": "Descarga los archivos XML y PDF de un CFDI existente",
                "produces": [
                    "application/pdf",
                    " application/xml"
                ],
                "tags": [
                    "CFDI"
                ],
                "summary": "Descargar PDF y XML de un CFDI",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del CFDI en Facturama",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Formato: xml o pdf",
                        "name": "format",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tipo de CFDI (issued)",
                        "name": "type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/sandbox/cfdi": {
            "post": {
                "description": "Genera una nueva factura CFDI (global o normal) en el ambiente de prueba",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CFDI"
                ],
                "summary": "Crear CFDI de prueba",
                "parameters": [
                    {
                        "description": "Datos del CFDI (estructura completa)",
                        "name": "factura",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CfdiRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.CfdiRequest": {
            "type": "object",
            "properties": {
                "CfdiType": {
                    "type": "string",
                    "example": "I"
                },
                "ExpeditionPlace": {
                    "type": "string",
                    "example": "20160"
                },
                "GlobalInformation": {
                    "$ref": "#/definitions/models.GlobalInformation"
                },
                "Items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Item"
                    }
                },
                "PaymentForm": {
                    "type": "string",
                    "example": "01"
                },
                "PaymentMethod": {
                    "type": "string",
                    "example": "PUE"
                },
                "Receiver": {
                    "$ref": "#/definitions/models.Receiver"
                }
            }
        },
        "models.GlobalInformation": {
            "type": "object",
            "properties": {
                "Months": {
                    "type": "string",
                    "example": "07"
                },
                "Periodicity": {
                    "description": "Mensual",
                    "type": "string",
                    "example": "04"
                },
                "Year": {
                    "type": "integer",
                    "example": 2025
                }
            }
        },
        "models.Item": {
            "type": "object",
            "properties": {
                "Description": {
                    "type": "string",
                    "example": "Ventas mes de Julio"
                },
                "ProductCode": {
                    "type": "string",
                    "example": "31162800"
                },
                "Quantity": {
                    "type": "number",
                    "example": 1
                },
                "Subtotal": {
                    "type": "number",
                    "example": 8767.24
                },
                "TaxObject": {
                    "type": "string",
                    "example": "02"
                },
                "Taxes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Tax"
                    }
                },
                "Total": {
                    "type": "number",
                    "example": 10170
                },
                "Unit": {
                    "type": "string",
                    "example": "Variedad"
                },
                "UnitCode": {
                    "type": "string",
                    "example": "AS"
                },
                "UnitPrice": {
                    "type": "number",
                    "example": 8767.24
                }
            }
        },
        "models.Receiver": {
            "type": "object",
            "properties": {
                "CfdiUse": {
                    "type": "string",
                    "example": "S01"
                },
                "FiscalRegime": {
                    "type": "string",
                    "example": "616"
                },
                "Name": {
                    "type": "string",
                    "example": "PUBLICO EN GENERAL"
                },
                "Rfc": {
                    "type": "string",
                    "example": "XAXX010101000"
                },
                "TaxZipCode": {
                    "type": "string",
                    "example": "20160"
                }
            }
        },
        "models.Tax": {
            "type": "object",
            "properties": {
                "Base": {
                    "type": "number",
                    "example": 8767.24
                },
                "IsFederalTax": {
                    "type": "boolean",
                    "example": true
                },
                "IsRetention": {
                    "type": "boolean",
                    "example": false
                },
                "Name": {
                    "type": "string",
                    "example": "IVA"
                },
                "Rate": {
                    "type": "number",
                    "example": 0.16
                },
                "Total": {
                    "type": "number",
                    "example": 1402.76
                }
            }
        }
    }
}